pyproject.toml
[project]
name = "qwc-oblique-imagery-service"
version = "v2025.11.27"
description = "QWC oblique imagery service"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "Flask~=3.0.0",
    "Werkzeug~=3.0.1",
    "Flask-JWT-Extended~=4.6.0",
    "flask-restx~=1.3.0",
    "qwc-services-core~=1.4.0",
    "rasterio>=1.4.3",
    "pillow>=12.0.0",
    "numpy>=2.2.6",
]

[dependency-groups]
dev = [
    "flask-cors>=5.0.0",
    "python-dotenv>=1.0.1",
]
